+++ CHANGELOG.md
@@ -3,6 +3,8 @@
 <!-- There is always Unreleased section on the top. Subsections (Added, Changed, Fixed, Removed) should be added as needed. -->
 
 ## Unreleased
+
+## 4.1.0 - 2021-04-14
 - Add SourceHut detection support.
 
 ## 4.0.0 - 2021-02-02